move sample rom file into util/ folder - backport change from v3.0
