+++ Diz.Test/Tests/LoadSaveTest.cs
@@ -5,6 +5,7 @@ using Diz.Core.model;
 using Diz.Core.serialization;
 using Diz.Core.serialization.xml_serializer;
 using Diz.Core.util;
+using Diz.Test.Tests;
 using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
